refactor: changed typeDefs directory to type-defs
