+++ koa-crud/src/graphql/type-defs/promos.ts
